Add tests for refactor of email
